+++ object/preprocessing/abstract_preprocessor.py
@@ -6,7 +6,7 @@ from abc import ABC, abstractmethod
 class AbstractPreprocessor(ABC):
 
     @abstractmethod
-    def process(self, frame: Array[np.int]):
+    def process(self, frame: Array[np.uint8]):
         """
         :param frame:
         :return processed frame or None if the frame is to be skipped:
